requirements.txt
# Core data processing
pandas>=1.3.0
numpy>=1.20.0

# Reddit data scraping
praw>=7.5.0

# Database connectivity
mysql-connector-python>=8.0.0

# Embeddings and vector similarity
sentence-transformers>=2.2.0

# Weaviate vector database
weaviate-client>=4.0.0

# Topic modeling (Step 1)
bertopic>=0.15.0
umap-learn>=0.5.0
hdbscan>=0.8.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Streamlit frontend
streamlit>=1.28.0

# HTTP requests for Ollama
requests>=2.25.0

# Progress bars
tqdm>=4.62.0

# Environment variables (optional)
python-dotenv>=0.19.0